migration-wrapper: Match ssh tunnel script more loosely
Currently the ssh tunnel script used by libvirt to connect to the
host is matched using pure string comparison, which works fine
but is a bit fragile in that even minor, semantically-irrelevant
changes to libvirt might break the detection logic: this has
happened in the past, when virt-ssh-helper was introduced and the
ssh tunnel script had to be modified accordingly.

Use regex matching instead. This way, we can focus on just the
relevant bits (is virt-ssh-helper or netcat being used?) and
ignore the irrelevant ones (how many spaces appear in between
tokens?).

For parts of the command line that have historically been enclosed
in quotes but are not actually required to be from the shell's
point of view (netcat binary name and libvirt connection URI),
both the unquoted and quoted versions are accepted to guarantee
compatibility with existing libvirt versions as well as upcoming
versions that might do away with the unnecessary quoting.

Note that this change doesn't in any way negatively impact
security: a rogue ssh client could have simply used the exact
same tunnel script as libvirt, and the migration wrapper would
have had no way to tell it apart from a real libvirtd instance.

  TODO
    update reference URLs

Signed-off-by: Andrea Bolognani <abologna@redhat.com>